created history schema
