+++ server/package.json
@@ -15,7 +15,7 @@
     "apollo-server-express": "^3.6.1",
     "bcrypt": "^5.0.1",
     "express": "^4.17.2",
-    "faker": "^6.6.6",
+    "faker": "^4.1.0",
     "graphql": "^16.2.0",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.6",
